update template, create additional dependancy for the route attachment
turns out in some cases there's a race condition upon creation of the
vpc default route and the vpc gateway; if the gateway hasn't been
created yet, the creation of the route fails.

this should now be fixed; the route creation should now wait for the
gateway to be created first.